Fold InvHelper into InventoryHelper
